Update common files to build in a separate module
