+++ Sources/FluentUI_iOS/Components/Bottom Commanding/BottomCommandingController.swift
@@ -60,7 +60,7 @@ public protocol BottomCommandingControllerDelegate: AnyObject {
 /// Items from the `expandedListSections` are either presented in an expanded sheet or a popover, depending on the current style.
 ///
 @objc(MSFBottomCommandingController)
-open class BottomCommandingController: UIViewController, TokenizedControlInternal {
+open class BottomCommandingController: UIViewController, TokenizedControl {
 
     /// View controller that will be displayed below the bottom commanding UI.
     @objc public var contentViewController: UIViewController? {
