+++ Sources/FluentUI_iOS/Components/Calendar/Views/CalendarViewDayCell.swift
@@ -50,7 +50,7 @@ let calendarViewDayCellVisualStateTransitionDuration: TimeInterval = 0.3
 
 // MARK: - CalendarViewDayCell
 
-class CalendarViewDayCell: UICollectionViewCell, TokenizedControlInternal {
+class CalendarViewDayCell: UICollectionViewCell, TokenizedControl {
     struct Constants {
         static let borderWidth: CGFloat = 0.5
         static let dotDiameter: CGFloat = 6.0
