+++ Sources/FluentUI_iOS/Components/Command Bar/CommandBar.swift
@@ -18,7 +18,7 @@ public protocol CommandBarDelegate: AnyObject {
  Provide `itemGroups` in `init` to set the buttons in the CommandBar. Optional `leadingItemGroups` and `trailingItemGroups` add buttons in leading and trailing positions. Each `CommandBarItem` will be represented as a button.
  */
 @objc(MSFCommandBar)
-public class CommandBar: UIView, TokenizedControlInternal {
+public class CommandBar: UIView, TokenizedControl {
     // Hierarchy:
     //
     // isScrollable = true
