+++ Sources/FluentUI_iOS/Components/Drawer/DrawerController.swift
@@ -97,7 +97,7 @@ public protocol DrawerControllerDelegate: AnyObject {
  */
 
 @objc(MSFDrawerController)
-open class DrawerController: UIViewController, TokenizedControlInternal {
+open class DrawerController: UIViewController, TokenizedControl {
     /// DrawerController colors with obj-c support
     @objc public static func drawerBackgroundColor(fluentTheme: FluentTheme?) -> UIColor {
         let theme = fluentTheme ?? .shared
