+++ Sources/FluentUI_iOS/Components/Label/BadgeLabel.swift
@@ -7,7 +7,7 @@ import UIKit
 
 // MARK: BadgeLabel
 
-class BadgeLabel: UILabel, TokenizedControlInternal {
+class BadgeLabel: UILabel, TokenizedControl {
     var style: BadgeLabelStyle = .system {
         didSet {
             updateColors()
