+++ Sources/FluentUI_iOS/Components/Label/Label.swift
@@ -9,7 +9,7 @@ import UIKit
 
 /// By default, `adjustsFontForContentSizeCategory` is set to true to automatically update its font when device's content size category changes
 @objc(MSFLabel)
-open class Label: UILabel, TokenizedControlInternal {
+open class Label: UILabel, TokenizedControl {
     private static let defaultColorForTheme: (FluentTheme) -> UIColor = TextColorStyle.regular.uiColor
 
     @objc open var colorStyle: TextColorStyle {
