+++ Sources/FluentUI_iOS/Components/Navigation/SearchBar/SearchBar.swift
@@ -22,7 +22,7 @@ public protocol SearchBarDelegate: AnyObject {
 
 /// Drop-in replacement for UISearchBar that allows for more customization
 @objc(MSFSearchBar)
-open class SearchBar: UIView, TokenizedControlInternal {
+open class SearchBar: UIView, TokenizedControl {
     @objc open var hidesNavigationBarDuringSearch: Bool = true {
         didSet {
             if oldValue != hidesNavigationBarDuringSearch && isActive {
