+++ Sources/FluentUI_iOS/Components/Navigation/Shy Header/ShyHeaderView.swift
@@ -11,7 +11,7 @@ import UIKit
 /// Used to contain an accessory provided by the VC contained by the NavigatableShyContainerVC
 /// This class in itself is fairly straightforward, defining a height and a containment layout
 /// The animation around showing/hiding this view progressively is handled by its superview/superVC, an instance of ShyHeaderController
-class ShyHeaderView: UIView, TokenizedControlInternal {
+class ShyHeaderView: UIView, TokenizedControl {
     typealias TokenSetKeyType = EmptyTokenSet.Tokens
     public var tokenSet: EmptyTokenSet = .init()
