+++ Sources/FluentUI_iOS/Components/Other Cells/ActivityIndicatorCell.swift
@@ -8,7 +8,7 @@ import UIKit
 // MARK: ActivityIndicatorCell
 
 @objc(MSFActivityIndicatorCell)
-open class ActivityIndicatorCell: UITableViewCell, TokenizedControlInternal {
+open class ActivityIndicatorCell: UITableViewCell, TokenizedControl {
     public static let identifier: String = "ActivityIndicatorCell"
 
     @objc public var backgroundStyleType: TableViewCellBackgroundStyleType = .plain {
