+++ Sources/FluentUI_iOS/Components/Other Cells/CenteredLabelCell.swift
@@ -8,7 +8,7 @@ import UIKit
 // MARK: CenteredLabelCell
 
 @objc(MSFCenteredLabelCell)
-open class CenteredLabelCell: UITableViewCell, TokenizedControlInternal {
+open class CenteredLabelCell: UITableViewCell, TokenizedControl {
     public static let identifier: String = "CenteredLabelCell"
 
     public typealias TokenSetKeyType = TableViewCellTokenSet.Tokens
