+++ Sources/FluentUI_iOS/Components/ResizingHandleView/ResizingHandleView.swift
@@ -8,7 +8,7 @@ import UIKit
 // MARK: - ResizingHandleView
 
 @objc(MSFResizingHandleView)
-open class ResizingHandleView: UIView, TokenizedControlInternal {
+open class ResizingHandleView: UIView, TokenizedControl {
     @objc public static let height: CGFloat = 20
 
     private lazy var markLayer: CALayer = {
