+++ Sources/FluentUI_iOS/Components/SegmentedControl/SegmentedControl.swift
@@ -7,7 +7,7 @@ import UIKit
 // MARK: SegmentedControl
 /// A styled segmented control that should be used instead of UISegmentedControl. It is designed to flex the button width proportionally to the control's width.
 @objc(MSFSegmentedControl)
-open class SegmentedControl: UIView, TokenizedControlInternal {
+open class SegmentedControl: UIView, TokenizedControl {
     private struct Constants {
         static let iPadMinimumWidth: CGFloat = 375
     }
