+++ Sources/FluentUI_iOS/Components/Separator/Separator.swift
@@ -16,7 +16,7 @@ public enum SeparatorOrientation: Int {
 // MARK: - Separator
 
 @objc(MSFSeparator)
-open class Separator: UIView, TokenizedControlInternal {
+open class Separator: UIView, TokenizedControl {
     public typealias TokenSetKeyType = SeparatorTokenSet.Tokens
     lazy public var tokenSet: SeparatorTokenSet = .init()
