+++ Sources/FluentUI_iOS/Components/Shimmer/ShimmerView.swift
@@ -7,7 +7,7 @@ import UIKit
 
 /// View that converts the subviews of a container view into a loading state with the "shimmering" effect.
 @objc(MSFShimmerView)
-open class ShimmerView: UIView, TokenizedControlInternal {
+open class ShimmerView: UIView, TokenizedControl {
 
     /// Optional synchronizer to sync multiple shimmer views.
     @objc open weak var animationSynchronizer: AnimationSynchronizerProtocol?
