+++ Sources/FluentUI_iOS/Components/Tab Bar/SideTabBar.swift
@@ -23,7 +23,7 @@ public protocol SideTabBarDelegate {
 /// View for a vertical side tab bar that can be used for app navigation.
 /// Optimized for horizontal regular + vertical regular size class configuration. Prefer using TabBarView for other size class configurations.
 @objc(MSFSideTabBar)
-open class SideTabBar: UIView, TokenizedControlInternal {
+open class SideTabBar: UIView, TokenizedControl {
     /// Delegate to handle user interactions in the side tab bar.
     @objc public weak var delegate: SideTabBarDelegate? {
         didSet {
