+++ Sources/FluentUI_iOS/Core/Theme/Tokens/ControlTokenSet.swift
@@ -4,11 +4,13 @@
 //
 
 import Combine
-import UIKit
 import SwiftUI
+#if os(iOS) || os(visionOS)
+import UIKit
+#endif // os(iOS) || os(visionOS)
 
 /// Base class for all Fluent control tokenization.
-public class ControlTokenSet<T: TokenSetKey>: ObservableObject {
+open class ControlTokenSet<T: TokenSetKey>: ObservableObject {
     /// Allows us to index into this token set using square brackets.
     ///
     /// We can use square brackets to both read and write into this `TokenSet`. For example:
@@ -61,7 +63,7 @@ public class ControlTokenSet<T: TokenSetKey>: ObservableObject {
     /// - Parameter otherTokenSet: The token set we will be pulling values from.
     /// - Parameter mapping: A `Dictionary` that maps our own tokens that we wish to override with
     /// their corresponding tokens in `otherTokenSet`.
-    func setOverrides<U>(from otherTokenSet: ControlTokenSet<U>, mapping: [T: U]) {
+    public func setOverrides<U>(from otherTokenSet: ControlTokenSet<U>, mapping: [T: U]) {
         // Make a copy so we write all the values at once
         var valueOverrideCopy = valueOverrides ?? [:]
         mapping.forEach { (thisToken, otherToken) in
@@ -71,7 +73,7 @@ public class ControlTokenSet<T: TokenSetKey>: ObservableObject {
     }
 
     /// Initialize the `ControlTokenSet` with an escaping callback for fetching default values.
-    init(_ defaults: @escaping (_ token: T, _ theme: FluentTheme) -> ControlTokenValue) {
+    public init(_ defaults: @escaping (_ token: T, _ theme: FluentTheme) -> ControlTokenValue) {
         self.defaults = defaults
     }
 
@@ -81,7 +83,7 @@ public class ControlTokenSet<T: TokenSetKey>: ObservableObject {
 
     /// Removes all `onUpdate`-based observing. Useful if you are re-registering the same tokenSet
     /// for a new instance of a control (see `Tooltip` for an example).
-    func deregisterOnUpdate() {
+    public func deregisterOnUpdate() {
         if let notificationObserver {
             NotificationCenter.default.removeObserver(notificationObserver,
                                                       name: .didChangeTheme,
@@ -95,7 +97,7 @@ public class ControlTokenSet<T: TokenSetKey>: ObservableObject {
     /// Prepares this token set by installing the current `FluentTheme` if it has changed.
     ///
     /// - Parameter fluentTheme: The current `FluentTheme` for the control's environment.
-    func update(_ fluentTheme: FluentTheme) {
+    public func update(_ fluentTheme: FluentTheme) {
         if fluentTheme != self.fluentTheme {
             self.fluentTheme = fluentTheme
         }
@@ -111,7 +113,7 @@ public class ControlTokenSet<T: TokenSetKey>: ObservableObject {
     /// - Parameter token: The token key to fetch any existing override for.
     ///
     /// - Returns: the active override value for a given token, or nil if none exists.
-    func overrideValue(forToken token: T) -> ControlTokenValue? {
+    public func overrideValue(forToken token: T) -> ControlTokenValue? {
         if let value = valueOverrides?[token] {
             return value
         } else if let value = fluentTheme.tokens(for: type(of: self))?[token] {
@@ -124,7 +126,7 @@ public class ControlTokenSet<T: TokenSetKey>: ObservableObject {
     ///
     /// - Parameter value: The value to set as an override.
     /// - Parameter token: The token key whose value should be set.
-    func setOverrideValue(_ value: ControlTokenValue?, forToken token: T) {
+    public func setOverrideValue(_ value: ControlTokenValue?, forToken token: T) {
         if valueOverrides == nil {
             valueOverrides = [:]
         }
@@ -142,7 +144,7 @@ public class ControlTokenSet<T: TokenSetKey>: ObservableObject {
     ///
     /// - Parameter control: The `UIView` instance that wishes to observe.
     /// - Parameter onUpdate: A callback to run whenever `control` should update itself.
-    func registerOnUpdate(for control: UIView, onUpdate: @escaping (() -> Void)) {
+    public func registerOnUpdate(for control: FluentThemeable, onUpdate: @escaping (() -> Void)) {
         guard self.onUpdate == nil,
               changeSink == nil,
               notificationObserver == nil else {
@@ -164,7 +166,7 @@ public class ControlTokenSet<T: TokenSetKey>: ObservableObject {
                                                                       queue: nil) { [weak self, weak control] notification in
             guard let strongSelf = self,
                   let control,
-                  FluentTheme.isApplicableThemeChange(notification, for: control)
+                  control.isApplicableThemeChange(notification)
             else {
                 return
             }
@@ -173,7 +175,7 @@ public class ControlTokenSet<T: TokenSetKey>: ObservableObject {
     }
 
     /// The current `FluentTheme` associated with this `ControlTokenSet`.
-    var fluentTheme: FluentTheme = FluentTheme.shared {
+    public var fluentTheme: FluentTheme = FluentTheme.shared {
         didSet {
             guard let onUpdate else {
                 return
@@ -204,9 +206,11 @@ public class ControlTokenSet<T: TokenSetKey>: ObservableObject {
 /// Union-type enumeration of all possible token values to be stored by a `ControlTokenSet`.
 public enum ControlTokenValue {
     case float(() -> CGFloat)
-    case uiColor(() -> UIColor)
     case color(() -> Color)
+#if os(iOS) || os(visionOS)
+    case uiColor(() -> UIColor)
     case uiFont(() -> UIFont)
+#endif // os(iOS) || os(visionOS)
     case shadowInfo(() -> ShadowInfo)
 
     public var float: CGFloat {
@@ -218,6 +222,21 @@ public enum ControlTokenValue {
         }
     }
 
+    public var color: Color {
+        if case .color(let color) = self {
+            return color()
+        } else {
+#if os(iOS) || os(visionOS)
+            if case .uiColor(let uiColor) = self {
+                return Color(uiColor())
+            }
+#endif // os(iOS) || os(visionOS)
+            assertionFailure("Cannot convert token to Color: \(self)")
+            return fallbackColor
+        }
+    }
+
+#if os(iOS) || os(visionOS)
     public var uiColor: UIColor {
         if case .uiColor(let uiColor) = self {
             return uiColor()
@@ -229,17 +248,6 @@ public enum ControlTokenValue {
         }
     }
 
-    public var color: Color {
-        if case .color(let color) = self {
-            return color()
-        } else if case .uiColor(let uiColor) = self {
-            return Color(uiColor())
-        } else {
-            assertionFailure("Cannot convert token to Color: \(self)")
-            return fallbackColor
-        }
-    }
-
     public var uiFont: UIFont {
         if case .uiFont(let uiFont) = self {
             return uiFont()
@@ -248,17 +256,18 @@ public enum ControlTokenValue {
             return UIFont()
         }
     }
+#endif // os(iOS) || os(visionOS)
 
     public var shadowInfo: ShadowInfo {
         if case .shadowInfo(let shadowInfo) = self {
             return shadowInfo()
         } else {
             assertionFailure("Cannot convert token to ShadowInfo: \(self)")
-            return ShadowInfo(keyColor: fallbackUIColor,
+            return ShadowInfo(keyColor: fallbackColor,
                               keyBlur: 10.0,
                               xKey: 10.0,
                               yKey: 10.0,
-                              ambientColor: fallbackUIColor,
+                              ambientColor: fallbackColor,
                               ambientBlur: 10.0,
                               xAmbient: 10.0,
                               yAmbient: 10.0)
@@ -284,10 +293,14 @@ public enum ControlTokenValue {
         switch value {
         case let number as NSNumber:
             self = .float { CGFloat(number.doubleValue) }
-        case let color as UIColor:
-            self = .uiColor { color }
+        case let color as Color:
+            self = .color { color }
+#if os(iOS) || os(visionOS)
+        case let uiColor as UIColor:
+            self = .uiColor { uiColor }
         case let font as UIFont:
             self = .uiFont { font }
+#endif // os(iOS) || os(visionOS)
         case let shadowInfo as ShadowInfo:
             self = .shadowInfo { shadowInfo }
         default:
@@ -297,14 +310,16 @@ public enum ControlTokenValue {
 
     // MARK: - Helpers
 
+#if os(iOS) || os(visionOS)
     private var fallbackUIColor: UIColor {
 #if DEBUG
         // Use our global "Hot Pink" in debug builds, to help identify unintentional conversions.
-        return GlobalTokens.sharedColor(.hotPink, .primary)
+        return UIColor(GlobalTokens.sharedSwiftUIColor(.hotPink, .primary))
 #else
         return GlobalTokens.neutralColor(.black)
 #endif
     }
+#endif // os(iOS) || os(visionOS)
 
     private var fallbackColor: Color {
 #if DEBUG
@@ -323,12 +338,14 @@ extension ControlTokenValue: CustomStringConvertible {
         switch self {
         case .float(let float):
             return "ControlTokenValue.float (\(float())"
-        case .uiColor(let uiColor):
-            return "ControlTokenValue.uiColor (\(uiColor())"
         case .color(let color):
             return "ControlTokenValue.color (\(color())"
+#if os(iOS) || os(visionOS)
+        case .uiColor(let uiColor):
+            return "ControlTokenValue.uiColor (\(uiColor())"
         case .uiFont(let uiFont):
             return "ControlTokenValue.uiFont (\(uiFont())"
+#endif // os(iOS) || os(visionOS)
         case .shadowInfo(let shadowInfo):
             return "ControlTokenValue.shadowInfo (\(shadowInfo())"
         }
