+++ Sources/FluentUI_iOS/Core/Theme/Tokens/ShadowInfo.swift
@@ -4,7 +4,8 @@
 //
 
 import CoreGraphics
-import UIKit
+import QuartzCore
+import SwiftUI
 
 /// Represents a two-part shadow as used by FluentUI.
 @objc(MSFShadowInfo)
@@ -20,11 +21,11 @@ public class ShadowInfo: NSObject {
     ///   - ambientBlur: The blur of the ambient shadow.
     ///   - xAmbient: The horizontal offset of the ambient shadow.
     ///   - yAmbient: The vertical offset of the ambient shadow.
-    public init(keyColor: UIColor,
+    public init(keyColor: Color,
                 keyBlur: CGFloat,
                 xKey: CGFloat,
                 yKey: CGFloat,
-                ambientColor: UIColor,
+                ambientColor: Color,
                 ambientBlur: CGFloat,
                 xAmbient: CGFloat,
                 yAmbient: CGFloat) {
@@ -39,7 +40,7 @@ public class ShadowInfo: NSObject {
     }
 
     /// The color of the key shadow.
-    @objc public let keyColor: UIColor
+    public let keyColor: Color
 
     /// The blur of the key shadow.
     @objc public let keyBlur: CGFloat
@@ -51,7 +52,7 @@ public class ShadowInfo: NSObject {
     @objc public let yKey: CGFloat
 
     /// The color of the ambient shadow.
-    @objc public let ambientColor: UIColor
+    public let ambientColor: Color
 
     /// The blur of the ambient shadow.
     @objc public let ambientBlur: CGFloat
@@ -66,55 +67,7 @@ public class ShadowInfo: NSObject {
     private let shadowBlurAdjustment: CGFloat = 0.5
 }
 
-public extension ShadowInfo {
-    /// Applies a key and an ambient shadow on a `UIView`.
-    /// - Parameters:
-    ///   - view: The view on which the shadows will be applied.
-    ///   - parentController: The view controller responsible for the view on which the shadows will be applied.
-    func applyShadow(to view: UIView, parentController: UIViewController? = nil) {
-        guard var shadowable = (view as? Shadowable) ?? (view.superview as? Shadowable) ?? (parentController as? Shadowable) else {
-            assertionFailure("Cannot apply Fluent shadows to a non-Shadowable view")
-            return
-        }
-
-        shadowable.ambientShadow?.removeFromSuperlayer()
-        shadowable.keyShadow?.removeFromSuperlayer()
-
-        let ambientShadow = initializeShadowLayer(view: view, isAmbientShadow: true)
-        let keyShadow = initializeShadowLayer(view: view)
-
-        shadowable.ambientShadow = ambientShadow
-        shadowable.keyShadow = keyShadow
-
-        view.layer.insertSublayer(ambientShadow, at: 0)
-        view.layer.insertSublayer(keyShadow, below: ambientShadow)
-    }
-
-    /// Initializes a `CALayer` with the relevant `ShadowInfo` values.
-    /// - Parameters:
-    ///   - view: The view on which the shadow layer will be applied.
-    ///   - isAmbientShadow: Determines whether to apply ambient or key shadow values on the layer.
-    /// - Returns: The modified `CALayer`.
-    func initializeShadowLayer(view: UIView, isAmbientShadow: Bool = false) -> CALayer {
-        let layer = CALayer()
-
-        layer.frame = view.bounds
-        layer.shadowColor = (isAmbientShadow ? ambientColor : keyColor).cgColor
-        layer.shadowRadius = isAmbientShadow ? ambientBlur : keyBlur
-
-        // The shadowOpacity needs to be set to 1 since the alpha is already set through shadowColor
-        layer.shadowOpacity = 1
-        layer.shadowOffset = CGSize(width: isAmbientShadow ? xAmbient : xKey,
-                                    height: isAmbientShadow ? yAmbient : yKey)
-        layer.needsDisplayOnBoundsChange = true
-        layer.cornerRadius = view.layer.cornerRadius
-        layer.backgroundColor = view.backgroundColor?.cgColor
-
-        return layer
-    }
-}
-
-/// Public protocol that, when implemented, allows any UIView or one of its subviews to implement fluent shadows
+/// Public protocol that, when implemented, allows any view or one of its subviews to implement fluent shadows
 public protocol Shadowable {
     /// The layer on which the ambient shadow is implemented
     var ambientShadow: CALayer? { get set }
