+++ Sources/FluentUI_macOS/Components/AvatarView/AvatarView.swift
@@ -637,66 +637,66 @@ extension AvatarView {
 
 	/// Table of background and text colors for the AvatarView
 	static let avatarColors: [ColorSet] = [
-		ColorSet(background: DynamicColor(light: Colors.Palette.darkRedTint40.color, dark: Colors.Palette.darkRedShade30.color),
-				 foreground: DynamicColor(light: Colors.Palette.darkRedShade30.color, dark: Colors.Palette.darkRedTint40.color)),
-		ColorSet(background: DynamicColor(light: Colors.Palette.cranberryTint40.color, dark: Colors.Palette.cranberryShade30.color),
-				 foreground: DynamicColor(light: Colors.Palette.cranberryShade30.color, dark: Colors.Palette.cranberryTint40.color)),
-		ColorSet(background: DynamicColor(light: Colors.Palette.redTint40.color, dark: Colors.Palette.redShade30.color),
-				 foreground: DynamicColor(light: Colors.Palette.redShade30.color, dark: Colors.Palette.redTint40.color)),
-		ColorSet(background: DynamicColor(light: Colors.Palette.pumpkinTint40.color, dark: Colors.Palette.pumpkinShade30.color),
-				 foreground: DynamicColor(light: Colors.Palette.pumpkinShade30.color, dark: Colors.Palette.pumpkinTint40.color)),
-		ColorSet(background: DynamicColor(light: Colors.Palette.peachTint40.color, dark: Colors.Palette.peachShade30.color),
-				 foreground: DynamicColor(light: Colors.Palette.peachShade30.color, dark: Colors.Palette.peachTint40.color)),
-		ColorSet(background: DynamicColor(light: Colors.Palette.marigoldTint40.color, dark: Colors.Palette.marigoldShade30.color),
-				 foreground: DynamicColor(light: Colors.Palette.marigoldShade30.color, dark: Colors.Palette.marigoldTint40.color)),
-		ColorSet(background: DynamicColor(light: Colors.Palette.goldTint40.color, dark: Colors.Palette.goldShade30.color),
-				 foreground: DynamicColor(light: Colors.Palette.goldShade30.color, dark: Colors.Palette.goldTint40.color)),
-		ColorSet(background: DynamicColor(light: Colors.Palette.brassTint40.color, dark: Colors.Palette.brassShade30.color),
-				 foreground: DynamicColor(light: Colors.Palette.brassShade30.color, dark: Colors.Palette.brassTint40.color)),
-		ColorSet(background: DynamicColor(light: Colors.Palette.brownTint40.color, dark: Colors.Palette.brownShade30.color),
-				 foreground: DynamicColor(light: Colors.Palette.brownShade30.color, dark: Colors.Palette.brownTint40.color)),
-		ColorSet(background: DynamicColor(light: Colors.Palette.forestTint40.color, dark: Colors.Palette.forestShade30.color),
-				 foreground: DynamicColor(light: Colors.Palette.forestShade30.color, dark: Colors.Palette.forestTint40.color)),
-		ColorSet(background: DynamicColor(light: Colors.Palette.seafoamTint40.color, dark: Colors.Palette.seafoamShade30.color),
-				 foreground: DynamicColor(light: Colors.Palette.seafoamShade30.color, dark: Colors.Palette.seafoamTint40.color)),
-		ColorSet(background: DynamicColor(light: Colors.Palette.darkGreenTint40.color, dark: Colors.Palette.darkGreenShade30.color),
-				 foreground: DynamicColor(light: Colors.Palette.darkGreenShade30.color, dark: Colors.Palette.darkGreenTint40.color)),
-		ColorSet(background: DynamicColor(light: Colors.Palette.lightTealTint40.color, dark: Colors.Palette.lightTealShade30.color),
-				 foreground: DynamicColor(light: Colors.Palette.lightTealShade30.color, dark: Colors.Palette.lightTealTint40.color)),
-		ColorSet(background: DynamicColor(light: Colors.Palette.tealTint40.color, dark: Colors.Palette.tealShade30.color),
-				 foreground: DynamicColor(light: Colors.Palette.tealShade30.color, dark: Colors.Palette.tealTint40.color)),
-		ColorSet(background: DynamicColor(light: Colors.Palette.steelTint40.color, dark: Colors.Palette.steelShade30.color),
-				 foreground: DynamicColor(light: Colors.Palette.steelShade30.color, dark: Colors.Palette.steelTint40.color)),
-		ColorSet(background: DynamicColor(light: Colors.Palette.blueTint40.color, dark: Colors.Palette.blueShade30.color),
-				 foreground: DynamicColor(light: Colors.Palette.blueShade30.color, dark: Colors.Palette.blueTint40.color)),
-		ColorSet(background: DynamicColor(light: Colors.Palette.royalBlueTint40.color, dark: Colors.Palette.royalBlueShade30.color),
-				 foreground: DynamicColor(light: Colors.Palette.royalBlueShade30.color, dark: Colors.Palette.royalBlueTint40.color)),
-		ColorSet(background: DynamicColor(light: Colors.Palette.cornFlowerTint40.color, dark: Colors.Palette.cornFlowerShade30.color),
-				 foreground: DynamicColor(light: Colors.Palette.cornFlowerShade30.color, dark: Colors.Palette.cornFlowerTint40.color)),
-		ColorSet(background: DynamicColor(light: Colors.Palette.navyTint40.color, dark: Colors.Palette.navyShade30.color),
-				 foreground: DynamicColor(light: Colors.Palette.navyShade30.color, dark: Colors.Palette.navyTint40.color)),
-		ColorSet(background: DynamicColor(light: Colors.Palette.lavenderTint40.color, dark: Colors.Palette.lavenderShade30.color),
-				 foreground: DynamicColor(light: Colors.Palette.lavenderShade30.color, dark: Colors.Palette.lavenderTint40.color)),
-		ColorSet(background: DynamicColor(light: Colors.Palette.purpleTint40.color, dark: Colors.Palette.purpleShade30.color),
-				 foreground: DynamicColor(light: Colors.Palette.purpleShade30.color, dark: Colors.Palette.purpleTint40.color)),
-		ColorSet(background: DynamicColor(light: Colors.Palette.grapeTint40.color, dark: Colors.Palette.grapeShade30.color),
-				 foreground: DynamicColor(light: Colors.Palette.grapeShade30.color, dark: Colors.Palette.grapeTint40.color)),
-		ColorSet(background: DynamicColor(light: Colors.Palette.lilacTint40.color, dark: Colors.Palette.lilacShade30.color),
-				 foreground: DynamicColor(light: Colors.Palette.lilacShade30.color, dark: Colors.Palette.lilacTint40.color)),
-		ColorSet(background: DynamicColor(light: Colors.Palette.pinkTint40.color, dark: Colors.Palette.pinkShade30.color),
-				 foreground: DynamicColor(light: Colors.Palette.pinkShade30.color, dark: Colors.Palette.pinkTint40.color)),
-		ColorSet(background: DynamicColor(light: Colors.Palette.magentaTint40.color, dark: Colors.Palette.magentaShade30.color),
-				 foreground: DynamicColor(light: Colors.Palette.magentaShade30.color, dark: Colors.Palette.magentaTint40.color)),
-		ColorSet(background: DynamicColor(light: Colors.Palette.plumTint40.color, dark: Colors.Palette.plumShade30.color),
-				 foreground: DynamicColor(light: Colors.Palette.plumShade30.color, dark: Colors.Palette.plumTint40.color)),
-		ColorSet(background: DynamicColor(light: Colors.Palette.beigeTint40.color, dark: Colors.Palette.beigeShade30.color),
-				 foreground: DynamicColor(light: Colors.Palette.beigeShade30.color, dark: Colors.Palette.beigeTint40.color)),
-		ColorSet(background: DynamicColor(light: Colors.Palette.minkTint40.color, dark: Colors.Palette.minkShade30.color),
-				 foreground: DynamicColor(light: Colors.Palette.minkShade30.color, dark: Colors.Palette.minkTint40.color)),
-		ColorSet(background: DynamicColor(light: Colors.Palette.platinumTint40.color, dark: Colors.Palette.platinumShade30.color),
-				 foreground: DynamicColor(light: Colors.Palette.platinumShade30.color, dark: Colors.Palette.platinumTint40.color)),
-		ColorSet(background: DynamicColor(light: Colors.Palette.anchorTint40.color, dark: Colors.Palette.anchorShade30.color),
-				 foreground: DynamicColor(light: Colors.Palette.anchorShade30.color, dark: Colors.Palette.anchorTint40.color))
+		ColorSet(background: DynamicColor_DEPRECATED(light: Colors.Palette.darkRedTint40.color, dark: Colors.Palette.darkRedShade30.color),
+				 foreground: DynamicColor_DEPRECATED(light: Colors.Palette.darkRedShade30.color, dark: Colors.Palette.darkRedTint40.color)),
+		ColorSet(background: DynamicColor_DEPRECATED(light: Colors.Palette.cranberryTint40.color, dark: Colors.Palette.cranberryShade30.color),
+				 foreground: DynamicColor_DEPRECATED(light: Colors.Palette.cranberryShade30.color, dark: Colors.Palette.cranberryTint40.color)),
+		ColorSet(background: DynamicColor_DEPRECATED(light: Colors.Palette.redTint40.color, dark: Colors.Palette.redShade30.color),
+				 foreground: DynamicColor_DEPRECATED(light: Colors.Palette.redShade30.color, dark: Colors.Palette.redTint40.color)),
+		ColorSet(background: DynamicColor_DEPRECATED(light: Colors.Palette.pumpkinTint40.color, dark: Colors.Palette.pumpkinShade30.color),
+				 foreground: DynamicColor_DEPRECATED(light: Colors.Palette.pumpkinShade30.color, dark: Colors.Palette.pumpkinTint40.color)),
+		ColorSet(background: DynamicColor_DEPRECATED(light: Colors.Palette.peachTint40.color, dark: Colors.Palette.peachShade30.color),
+				 foreground: DynamicColor_DEPRECATED(light: Colors.Palette.peachShade30.color, dark: Colors.Palette.peachTint40.color)),
+		ColorSet(background: DynamicColor_DEPRECATED(light: Colors.Palette.marigoldTint40.color, dark: Colors.Palette.marigoldShade30.color),
+				 foreground: DynamicColor_DEPRECATED(light: Colors.Palette.marigoldShade30.color, dark: Colors.Palette.marigoldTint40.color)),
+		ColorSet(background: DynamicColor_DEPRECATED(light: Colors.Palette.goldTint40.color, dark: Colors.Palette.goldShade30.color),
+				 foreground: DynamicColor_DEPRECATED(light: Colors.Palette.goldShade30.color, dark: Colors.Palette.goldTint40.color)),
+		ColorSet(background: DynamicColor_DEPRECATED(light: Colors.Palette.brassTint40.color, dark: Colors.Palette.brassShade30.color),
+				 foreground: DynamicColor_DEPRECATED(light: Colors.Palette.brassShade30.color, dark: Colors.Palette.brassTint40.color)),
+		ColorSet(background: DynamicColor_DEPRECATED(light: Colors.Palette.brownTint40.color, dark: Colors.Palette.brownShade30.color),
+				 foreground: DynamicColor_DEPRECATED(light: Colors.Palette.brownShade30.color, dark: Colors.Palette.brownTint40.color)),
+		ColorSet(background: DynamicColor_DEPRECATED(light: Colors.Palette.forestTint40.color, dark: Colors.Palette.forestShade30.color),
+				 foreground: DynamicColor_DEPRECATED(light: Colors.Palette.forestShade30.color, dark: Colors.Palette.forestTint40.color)),
+		ColorSet(background: DynamicColor_DEPRECATED(light: Colors.Palette.seafoamTint40.color, dark: Colors.Palette.seafoamShade30.color),
+				 foreground: DynamicColor_DEPRECATED(light: Colors.Palette.seafoamShade30.color, dark: Colors.Palette.seafoamTint40.color)),
+		ColorSet(background: DynamicColor_DEPRECATED(light: Colors.Palette.darkGreenTint40.color, dark: Colors.Palette.darkGreenShade30.color),
+				 foreground: DynamicColor_DEPRECATED(light: Colors.Palette.darkGreenShade30.color, dark: Colors.Palette.darkGreenTint40.color)),
+		ColorSet(background: DynamicColor_DEPRECATED(light: Colors.Palette.lightTealTint40.color, dark: Colors.Palette.lightTealShade30.color),
+				 foreground: DynamicColor_DEPRECATED(light: Colors.Palette.lightTealShade30.color, dark: Colors.Palette.lightTealTint40.color)),
+		ColorSet(background: DynamicColor_DEPRECATED(light: Colors.Palette.tealTint40.color, dark: Colors.Palette.tealShade30.color),
+				 foreground: DynamicColor_DEPRECATED(light: Colors.Palette.tealShade30.color, dark: Colors.Palette.tealTint40.color)),
+		ColorSet(background: DynamicColor_DEPRECATED(light: Colors.Palette.steelTint40.color, dark: Colors.Palette.steelShade30.color),
+				 foreground: DynamicColor_DEPRECATED(light: Colors.Palette.steelShade30.color, dark: Colors.Palette.steelTint40.color)),
+		ColorSet(background: DynamicColor_DEPRECATED(light: Colors.Palette.blueTint40.color, dark: Colors.Palette.blueShade30.color),
+				 foreground: DynamicColor_DEPRECATED(light: Colors.Palette.blueShade30.color, dark: Colors.Palette.blueTint40.color)),
+		ColorSet(background: DynamicColor_DEPRECATED(light: Colors.Palette.royalBlueTint40.color, dark: Colors.Palette.royalBlueShade30.color),
+				 foreground: DynamicColor_DEPRECATED(light: Colors.Palette.royalBlueShade30.color, dark: Colors.Palette.royalBlueTint40.color)),
+		ColorSet(background: DynamicColor_DEPRECATED(light: Colors.Palette.cornFlowerTint40.color, dark: Colors.Palette.cornFlowerShade30.color),
+				 foreground: DynamicColor_DEPRECATED(light: Colors.Palette.cornFlowerShade30.color, dark: Colors.Palette.cornFlowerTint40.color)),
+		ColorSet(background: DynamicColor_DEPRECATED(light: Colors.Palette.navyTint40.color, dark: Colors.Palette.navyShade30.color),
+				 foreground: DynamicColor_DEPRECATED(light: Colors.Palette.navyShade30.color, dark: Colors.Palette.navyTint40.color)),
+		ColorSet(background: DynamicColor_DEPRECATED(light: Colors.Palette.lavenderTint40.color, dark: Colors.Palette.lavenderShade30.color),
+				 foreground: DynamicColor_DEPRECATED(light: Colors.Palette.lavenderShade30.color, dark: Colors.Palette.lavenderTint40.color)),
+		ColorSet(background: DynamicColor_DEPRECATED(light: Colors.Palette.purpleTint40.color, dark: Colors.Palette.purpleShade30.color),
+				 foreground: DynamicColor_DEPRECATED(light: Colors.Palette.purpleShade30.color, dark: Colors.Palette.purpleTint40.color)),
+		ColorSet(background: DynamicColor_DEPRECATED(light: Colors.Palette.grapeTint40.color, dark: Colors.Palette.grapeShade30.color),
+				 foreground: DynamicColor_DEPRECATED(light: Colors.Palette.grapeShade30.color, dark: Colors.Palette.grapeTint40.color)),
+		ColorSet(background: DynamicColor_DEPRECATED(light: Colors.Palette.lilacTint40.color, dark: Colors.Palette.lilacShade30.color),
+				 foreground: DynamicColor_DEPRECATED(light: Colors.Palette.lilacShade30.color, dark: Colors.Palette.lilacTint40.color)),
+		ColorSet(background: DynamicColor_DEPRECATED(light: Colors.Palette.pinkTint40.color, dark: Colors.Palette.pinkShade30.color),
+				 foreground: DynamicColor_DEPRECATED(light: Colors.Palette.pinkShade30.color, dark: Colors.Palette.pinkTint40.color)),
+		ColorSet(background: DynamicColor_DEPRECATED(light: Colors.Palette.magentaTint40.color, dark: Colors.Palette.magentaShade30.color),
+				 foreground: DynamicColor_DEPRECATED(light: Colors.Palette.magentaShade30.color, dark: Colors.Palette.magentaTint40.color)),
+		ColorSet(background: DynamicColor_DEPRECATED(light: Colors.Palette.plumTint40.color, dark: Colors.Palette.plumShade30.color),
+				 foreground: DynamicColor_DEPRECATED(light: Colors.Palette.plumShade30.color, dark: Colors.Palette.plumTint40.color)),
+		ColorSet(background: DynamicColor_DEPRECATED(light: Colors.Palette.beigeTint40.color, dark: Colors.Palette.beigeShade30.color),
+				 foreground: DynamicColor_DEPRECATED(light: Colors.Palette.beigeShade30.color, dark: Colors.Palette.beigeTint40.color)),
+		ColorSet(background: DynamicColor_DEPRECATED(light: Colors.Palette.minkTint40.color, dark: Colors.Palette.minkShade30.color),
+				 foreground: DynamicColor_DEPRECATED(light: Colors.Palette.minkShade30.color, dark: Colors.Palette.minkTint40.color)),
+		ColorSet(background: DynamicColor_DEPRECATED(light: Colors.Palette.platinumTint40.color, dark: Colors.Palette.platinumShade30.color),
+				 foreground: DynamicColor_DEPRECATED(light: Colors.Palette.platinumShade30.color, dark: Colors.Palette.platinumTint40.color)),
+		ColorSet(background: DynamicColor_DEPRECATED(light: Colors.Palette.anchorTint40.color, dark: Colors.Palette.anchorShade30.color),
+				 foreground: DynamicColor_DEPRECATED(light: Colors.Palette.anchorShade30.color, dark: Colors.Palette.anchorTint40.color))
 	]
 }
