+++ Sources/FluentUI_macOS/Components/DynamicColor/DynamicColor.swift
@@ -7,17 +7,17 @@ import AppKit
 
 @objc(MSFColorSet)
 public class ColorSet: NSObject {
-	@objc public let background: DynamicColor
-	@objc public let foreground: DynamicColor
+	@objc public let background: DynamicColor_DEPRECATED
+	@objc public let foreground: DynamicColor_DEPRECATED
 
-	public init(background: DynamicColor, foreground: DynamicColor) {
+	public init(background: DynamicColor_DEPRECATED, foreground: DynamicColor_DEPRECATED) {
 		self.background = background
 		self.foreground = foreground
 	}
 }
 
-@objc(MSFDynamicColor)
-public class DynamicColor: NSObject {
+@objc(MSFDynamicColor_DEPRECATED)
+public class DynamicColor_DEPRECATED: NSObject {
 
 	@objc public let light: NSColor
 	@objc public let dark: NSColor
@@ -36,7 +36,7 @@ public class DynamicColor: NSObject {
 	}
 
 	public override func isEqual(_ object: Any?) -> Bool {
-		guard let dynamicColor = object as? DynamicColor else {
+		guard let dynamicColor = object as? DynamicColor_DEPRECATED else {
 			return false
 		}
 		return dynamicColor.light.isEqual(self.light) && dynamicColor.dark.isEqual(self.dark)
